+++ FirebaseAuth/Sources/Swift/Utilities/AuthErrors.swift
@@ -346,6 +346,9 @@ import Foundation
   /// describing which step of the JWT parsing process failed.
   case malformedJWT = 18000
 
+  /// Indicates that the password does not meet the project's password policy.
+  case passwordDoesNotMeetRequirements = 18001
+
   var errorDescription: String {
     switch self {
     case .invalidCustomToken:
@@ -528,6 +531,8 @@ import Foundation
       return kErrorSiteKeyMissing
     case .recaptchaActionCreationFailed:
       return kErrorRecaptchaActionCreationFailed
+    case .passwordDoesNotMeetRequirements:
+      return kErrorPasswordDoesNotMeetRequirements
     }
   }
 
@@ -719,6 +724,8 @@ import Foundation
       return "ERROR_RECAPTCHA_SITE_KEY_MISSING"
     case .recaptchaActionCreationFailed:
       return "ERROR_RECAPTCHA_ACTION_CREATION_FAILED"
+    case .passwordDoesNotMeetRequirements:
+      return "ERROR_PASSWORD_DOES_NOT_MEET_REQUIREMENTS"
     }
   }
 }
@@ -996,3 +1003,7 @@ private let kErrorSiteKeyMissing =
 private let kErrorRecaptchaActionCreationFailed =
   "The reCAPTCHA SDK action class failed to initialize. See " +
   "https://cloud.google.com/recaptcha-enterprise/docs/instrument-ios-apps"
+
+private let kErrorPasswordDoesNotMeetRequirements =
+  "The password does not meet the requirements."
+