+++ FirebaseAuth/Tests/Unit/SwiftAPI.swift
@@ -299,6 +299,7 @@ class AuthAPI_hOnlyTests: XCTestCase {
     _ = AuthErrorCode.keychainError
     _ = AuthErrorCode.internalError
     _ = AuthErrorCode.malformedJWT
+    _ = AuthErrorCode.passwordDoesNotMeetRequirements
   }
 
   func authSettings(settings: AuthSettings) {